Create README.md
Both my partner and I are having communication issues after multiple of years of holding an understanding for one another. Currently we are realizing that both of us at times seem to be powerless. Its as if exterior forces and or energies are implementing their own agenda or
 schedule(s) and its putting the organic us at serious risk. I hold little fear but question the well being of both her and I. I am seeking further collection of information to better understand and govern such realities so I can better support her and preserve the better being of our future. 
if you are in a similar situation, please feel free to share and for all whom have read the above:
Thank you for your time and God Bless to all !